Initial attempt at a --convert option to convert HFS formatted ipods (Macpods) to FAT32 format.  This needs testing (as well as the existing --format feature) before adding as an option in the interactive mode and fully documenting.
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@14030 a1c6a512-1295-4272-9138-f99709370657

+++ rbutil/ipodpatcher/ipodio.h
@@ -53,9 +53,9 @@ struct ipod_directory_t {
 };
 
 struct partinfo_t {
-  unsigned long start; /* first sector (LBA) */
-  unsigned long size;  /* number of sectors */
-  int type;
+  uint32_t start; /* first sector (LBA) */
+  uint32_t size;  /* number of sectors */
+  uint32_t type;
 };
 
 struct ipod_t {
